Separate old xfcapi to help with migration
This will help us keep track of what is being used and what has been
migrated over.
